Updated DIO Entendendo Métodos JAVA -Calculadora alterado floats para doubles

+++ src/DIO/EntendendoMetodosJava/Calculadora.java
@@ -1,22 +1,22 @@
 package DIO.EntendendoMetodosJava;
 
 public class Calculadora {
-    private static float soma (float a, float b){
+    private static double soma (double a, double b){
         return a + b;
     }
-    private static float subtracao (float a, float b){
+    private static double subtracao (double a, double b){
         return a - b;
     }
-    private static float multiplicacao (float a, float b){
+    private static double multiplicacao (double a, double b){
         return a * b;
     }
-    private static float divisao (float a, float b){
+    private static double divisao (double a, double b){
         return a / b;
     }
 
     public static void main(String[] args) {
-        float a = 12.0f;
-        float b = 3.0f;
+        double a = 12.0f;
+        double b = 3.0f;
         System.out.println(soma(a, b));
         System.out.println(subtracao(a, b));
         System.out.println(multiplicacao(a, b));
